Refactor app into modules with tests and CI
