+++ requirements.txt
@@ -1,0 +1,5 @@
+streamlit
+Pillow
+cryptography
+fastapi
+pytest
